# docs-site/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Overlay API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/subscription/settings": {
      "get": {
        "operationId": "getSubscriptionSettings",
        "tags": [
          "subscription"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionSettingsResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "postSubscriptionSettings",
        "tags": [
          "subscription"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionSettingsResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/topups/checkout": {
      "post": {
        "operationId": "postTopupsCheckout",
        "tags": [
          "topups"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopupsCheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopupsCheckoutResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/topups/verify": {
      "post": {
        "operationId": "postTopupsVerify",
        "tags": [
          "topups"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopupsVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopupsVerifyResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SubscriptionSettingsRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "SubscriptionSettingsResponse": {
        "type": "object",
        "additionalProperties": true
      },
      "TopupsCheckoutRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "TopupsCheckoutResponse": {
        "type": "object",
        "additionalProperties": true
      },
...
      "TopupsVerifyRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "TopupsVerifyResponse": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  }
}